Rename selvpc package to the selvpcclient
Rename package to use the same name for repo and a package.
