+++ scripts/format
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Check if Go files have bad code format.
-gofmt_diff="$(gofmt -d -s ./selvpc)"
+gofmt_diff="$(gofmt -d -s ./selvpcclient)"
 if [[ -n "${gofmt_diff}" ]]; then
   echo "${gofmt_diff}"
   exit 1
