+++ scripts/imports
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Find all Go files and check if they have wrong format of imports.
-files="$(find ./selvpc -type f -name '*.go')"
+files="$(find ./selvpcclient -type f -name '*.go')"
 goimports_diff="$(goimports -l -d $files)"
 if [[ -n "${goimports_diff}" ]]; then
   echo "${goimports_diff}"
